Adicionando Float

+++ horadecodar2.c
@@ -1,24 +1,24 @@
 #include <stdio.h>
 int main() {
-    int primeira_nota, segunda_nota, terceira_nota;
-    int media;
+    float primeira_nota, segunda_nota, terceira_nota;
+    float media;
 
 
         printf("***Programa de Cálculo de Média*** \n");
 
 
         printf("Digite a sua primeira nota: \n");
-        scanf("%d", &primeira_nota);
+        scanf("%f", &primeira_nota);
 
         printf("Digite a sua segunda nota: \n");
-        scanf("%d", &segunda_nota);
+        scanf("%f", &segunda_nota);
 
         printf("Digite a sua terceira nota: \n");
-        scanf("%d", &terceira_nota);
+        scanf("%f", &terceira_nota);
 
         media = (primeira_nota + segunda_nota + terceira_nota) /3;
 
-        printf("A média das suas notas é: %d", media);
+        printf("A média das suas notas é: %.3f", media);
